+++ src/lib/components/MarkdownIt.svelte
@@ -31,7 +31,7 @@
 <style>
    div.markdown-it {
       color: white;
-      line-height: 1.3em;
+      /*line-height: 1.3em;*/
       text-align: left;
    }
 </style>